move TeacherCard to its own component
